requirements.txt
Falsk
Flask-JWT-Extended
Flask-RESTful
Flask-SQLAlchemy
